the calculation of frequency registers was now done in the microcontroller, input is frequency like 439.9875 for 439,9875 MHz

+++ Encoder AVR Code/SI4432.h
@@ -25,7 +25,7 @@
 #define GPIO_DATA_CLK PORTB0 //gpio1
 #define GPIO_DATA_IN PORTB1 //gpio0
 
-uint32_t xmitfreq;
+float xmitfreq;
 int16_t offset;
 
 void SIXmit();
